Allow the dispatcher to be freed once a request is finished.
Previously, if there was no ref to the dispatcher, it would be freed
before the request could be satisfied. By keeping a ref in the
DispatcherRequestData, it is freed at the right time. This also
disconnects the signal handler from all channels and connections
when freeing the dispatcher as the "invalidated" signal callback can be
called after the dispatcher has been freed.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>

git-svn-id: svn+ssh://svn.gnome.org/svn/empathy/trunk@2306 4ee84921-47dd-4033-b63a-18d7a039a3e4